requirements.txt
# Web Framework
flask==3.0.0
gunicorn==21.2.0

# Database
psycopg2-binary==2.9.9

# Configuration
python-dotenv==1.0.0

# Development
pytest==7.4.3
